requirements.txt
python-dotenv
google-generativeai
flask
bootstrap-flask
flask-wtf

